+++ src/app/common/apis/splatoon3.ink/types/bankara_properties.ts
