+++ src/app/common/apis/splatoon3.ink/types/event_properties.ts
