+++ src/app/common/apis/splatoon3.ink/types/fest_properties.ts
