+++ src/app/common/apis/splatoon3.ink/types/locale.ts
