+++ src/app/common/apis/splatoon3.ink/types/regular_properties.ts
