+++ src/app/common/apis/splatoon3.ink/types/salmon_properties.ts
