+++ src/app/common/apis/splatoon3.ink/types/schedule.ts
