+++ src/app/common/apis/splatoon3.ink/types/x_properties.ts
