+++ src/app/common/apis/voice-text/voice_text.ts
