+++ src/app/feat-recruit/canvases/big_run_canvas.ts
@@ -4,8 +4,8 @@ import Canvas from 'canvas';
 import { RecruitOpCode } from './regenerate_canvas.js';
 import { modalRecruit } from '../../../constant.js';
 import { log4js_obj } from '../../../log4js_settings';
-import { getBigRunData } from '../../common/apis/splatoon3_ink';
-import { Sp3Schedule } from '../../common/apis/types/schedule.js';
+import { getBigRunData } from '../../common/apis/splatoon3.ink/splatoon3_ink.js';
+import { Sp3Schedule } from '../../common/apis/splatoon3.ink/types/schedule.js';
 import { createRoundRect, drawArcImage, fillTextWithStroke } from '../../common/canvas_components';
 import { dateformat, formatDatetime } from '../../common/convert_datetime';
 import { exists, notExists } from '../../common/others.js';
