+++ src/app/feat-recruit/interactions/commands/fes_recruit.ts
@@ -3,7 +3,7 @@ import { AttachmentBuilder, ChatInputCommandInteraction, GuildMember, Permission
 import { ParticipantService } from '../../../../db/participant_service';
 import { RecruitService, RecruitType } from '../../../../db/recruit_service';
 import { log4js_obj } from '../../../../log4js_settings';
-import { checkFes, getSchedule, getFesData, MatchInfo } from '../../../common/apis/splatoon3_ink';
+import { checkFes, getSchedule, getFesData, MatchInfo } from '../../../common/apis/splatoon3.ink/splatoon3_ink';
 import { setButtonDisable } from '../../../common/button_components';
 import { getGuildByInteraction } from '../../../common/manager/guild_manager';
 import { searchAPIMemberById, searchDBMemberById } from '../../../common/manager/member_manager';
