+++ src/app/feat-recruit/interactions/commands/salmon_recruit.ts
@@ -3,8 +3,14 @@ import { AttachmentBuilder, ChatInputCommandInteraction, GuildMember, Permission
 import { ParticipantService } from '../../../../db/participant_service';
 import { RecruitService, RecruitType } from '../../../../db/recruit_service';
 import { log4js_obj } from '../../../../log4js_settings';
-import { checkBigRun, checkTeamContest, getSalmonData, getSchedule, getTeamContestData } from '../../../common/apis/splatoon3_ink';
-import { Sp3Schedule } from '../../../common/apis/types/schedule';
+import {
+    checkBigRun,
+    checkTeamContest,
+    getSalmonData,
+    getSchedule,
+    getTeamContestData,
+} from '../../../common/apis/splatoon3.ink/splatoon3_ink';
+import { Sp3Schedule } from '../../../common/apis/splatoon3.ink/types/schedule';
 import { getGuildByInteraction } from '../../../common/manager/guild_manager';
 import { searchAPIMemberById, searchDBMemberById } from '../../../common/manager/member_manager';
 import { searchMessageById } from '../../../common/manager/message_manager';
