+++ src/app/feat-recruit/interactions/modals/fes_recruit_modal.ts
@@ -4,7 +4,7 @@ import { AttachmentBuilder, ModalSubmitInteraction } from 'discord.js';
 import { ParticipantService } from '../../../../db/participant_service';
 import { RecruitService, RecruitType } from '../../../../db/recruit_service';
 import { log4js_obj } from '../../../../log4js_settings';
-import { MatchInfo } from '../../../common/apis/splatoon3_ink';
+import { MatchInfo } from '../../../common/apis/splatoon3.ink/splatoon3_ink';
 import { setButtonDisable } from '../../../common/button_components';
 import { getGuildByInteraction } from '../../../common/manager/guild_manager';
 import { searchMessageById } from '../../../common/manager/message_manager';
