+++ src/app/feat-utils/splat3/stageinfo.ts
@@ -9,8 +9,8 @@ import {
     getXMatchData,
     checkFes,
     getSchedule,
-} from '../../common/apis/splatoon3_ink';
-import { Sp3Schedule } from '../../common/apis/types/schedule';
+} from '../../common/apis/splatoon3.ink/splatoon3_ink';
+import { Sp3Schedule } from '../../common/apis/splatoon3.ink/types/schedule';
 import { formatDatetime, dateformat } from '../../common/convert_datetime.js';
 import { assertExistCheck } from '../../common/others';
